pdf2docx/requirements.txt
PyPDF2>=3.0.0
python-docx>=0.8.11
click>=8.0.0